requirements/base.txt
# Django
Django==5.1
djangorestframework==3.15.2
django-filter==24.3
django-environ==0.11.2

# Security
argon2-cffi==23.1.0

# Docs
drf-spectacular==0.27.2

# Utilities
djangorestframework-jsonapi==7.1.0
pillow==10.3.0
Markdown==3.6
